+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Index-based section extraction now correctlyignores headings inside code blocks (WI-2026-02-02-001)
+
 ## [0.2.0] - 2026-01-31
 
 ### Added